+++ packages/pangea-sdk/PangeaCyber.Net/AIGuard/Results/TextGuardResult.cs
@@ -8,15 +8,21 @@ namespace PangeaCyber.Net.AIGuard.Results;
 [JsonObject(NamingStrategyType = typeof(SnakeCaseNamingStrategy))]
 public sealed class TextGuardResult<TMessages>
 {
-    /// <summary>Result of the recipe analyzing and input prompt.</summary>
-    public TextGuardDetectors Detectors { get; set; } = default!;
-
     /// <summary>Updated prompt text, if applicable.</summary>
     public string? PromptText { get; set; }
 
     /// <summary>Updated structured prompt, if applicable.</summary>
     public TMessages? PromptMessages { get; set; }
 
-    /// <summary>Blocked</summary>
+    /// <summary>Whether or not the prompt triggered a block detection.</summary>
     public bool Blocked { get; set; }
+
+    /// <summary>The Recipe that was used.</summary>
+    public string Recipe { get; set; } = default!;
+
+    /// <summary>Result of the recipe analyzing and input prompt.</summary>
+    public TextGuardDetectors Detectors { get; set; } = default!;
+
+    /// <summary>If an FPE redaction method returned results, this will be the context passed to unredact.</summary>
+    public string? FpeContext { get; set; }
 }