+++ packages/pangea-sdk/CHANGELOG.md
@@ -12,12 +12,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Prompt Guard: `classify` parameter.
 - AI Guard: support for structured input.
 - AI Guard: more detectors.
+- AI Guard: support for `llm_input`, `llm_info`, and `log_fields`.
 
 ### Changed
 
 - Clarified what `Config.Environment` affects.
 - AI Guard: replaced `redacted` with `action`.
 - AI Guard: updated to v1.
+- Prompt Guard: updated to v1.
 
 ## 4.4.0-beta.2 - 2025-01-21